+++ msip-bhu-business-impl/msip_bhu_rpc_api/src/main/java/com/bhu/vas/api/rpc/daemon/helper/DaemonHelper.java
@@ -57,7 +57,7 @@ public class DaemonHelper {
 		daemonRpcService.wifiDeviceCmdDown(null, mac, cmd);
 	}
 	
-	public static void locationStep2Query(String mac,int taskid,String serialno,IDaemonRpcService daemonRpcService){
+	public static void locationStep2Query(String mac,long taskid,String serialno,IDaemonRpcService daemonRpcService){
 		String cmd = CMDBuilder.builderDeviceLocationStep2Query(mac, taskid, serialno);
 		//String cmd = CMDBuilder.builderDeviceLocationNotifyQuery(mac, CMDBuilder.location_taskid_fragment.getNextSequence());
 		daemonRpcService.wifiDeviceCmdDown(null, mac, cmd);
