+++ msip-bhu-business-impl/msip_bhu_rpc_api/src/main/java/com/bhu/vas/api/rpc/user/iservice/IUserCaptchaCodeRpcService.java
@@ -15,7 +15,7 @@ public interface IUserCaptchaCodeRpcService {
 			String channel_taskid);
 	
 	public void taskStatusFetch(
-			int taskid
+			long taskid
 			);
 	
 	public void taskStatusFetch4ThirdParties(
