+++ msip-bhu-business-impl/msip_bhu_rpc_api/src/main/java/com/bhu/vas/api/rpc/user/iservice/IUserRpcService.java
@@ -84,7 +84,7 @@ public interface IUserRpcService {
 			String channel_taskid);
 	
 	public void taskStatusFetch(
-			int taskid
+			long taskid
 			);
 	
 	public void taskStatusFetch4ThirdParties(
